fix(core): center peek and history dialog does not display custom fonts (#9274)
close AF-1806
