+++ packages/frontend/core/package.json
@@ -37,6 +37,7 @@
     "@radix-ui/react-dialog": "^1.1.3",
     "@radix-ui/react-popover": "^1.1.3",
     "@radix-ui/react-scroll-area": "^1.2.2",
+    "@radix-ui/react-slot": "^1.1.1",
     "@radix-ui/react-toolbar": "^1.1.1",
     "@sentry/react": "^8.44.0",
     "@toeverything/infra": "workspace:*",
