+++ packages/frontend/core/src/components/affine/page-history-modal/history-modal.tsx
@@ -30,7 +30,10 @@ import {
 import { encodeStateAsUpdate } from 'yjs';
 
 import { pageHistoryModalAtom } from '../../atoms/page-history';
-import { BlockSuiteEditor } from '../../blocksuite/block-suite-editor';
+import {
+  BlockSuiteEditor,
+  CustomEditorWrapper,
+} from '../../blocksuite/block-suite-editor';
 import { PureEditorModeSwitch } from '../../blocksuite/block-suite-mode-switch';
 import { AffineErrorBoundary } from '../affine-error-boundary';
 import {
@@ -128,11 +131,13 @@ const HistoryEditorPreview = ({
           <AffineErrorBoundary>
             <Scrollable.Root>
               <Scrollable.Viewport className="affine-page-viewport">
-                <BlockSuiteEditor
-                  className={styles.editor}
-                  mode={mode}
-                  page={snapshotPage}
-                />
+                <CustomEditorWrapper>
+                  <BlockSuiteEditor
+                    className={styles.editor}
+                    mode={mode}
+                    page={snapshotPage}
+                  />
+                </CustomEditorWrapper>
               </Scrollable.Viewport>
               <Scrollable.Scrollbar />
             </Scrollable.Root>
