+++ packages/frontend/core/src/components/blocksuite/block-suite-editor/index.ts
@@ -10,3 +10,4 @@ edgelessEffects();
 registerBlocksuitePresetsCustomComponents();
 
 export * from './blocksuite-editor';
+export * from './custom-editor-wrapper';
