+++ packages/frontend/core/src/components/page-detail-editor.tsx
@@ -2,18 +2,15 @@ import './page-detail-editor.css';
 
 import type { AffineEditorContainer } from '@blocksuite/affine/presets';
 import { useLiveData, useService } from '@toeverything/infra';
-import { cssVar } from '@toeverything/theme';
 import clsx from 'clsx';
-import type { CSSProperties } from 'react';
-import { useMemo } from 'react';
 
 import { DocService } from '../modules/doc';
 import { EditorService } from '../modules/editor';
+import { EditorSettingService } from '../modules/editor-setting';
 import {
-  EditorSettingService,
-  fontStyleOptions,
-} from '../modules/editor-setting';
-import { BlockSuiteEditor as Editor } from './blocksuite/block-suite-editor';
+  BlockSuiteEditor as Editor,
+  CustomEditorWrapper,
+} from './blocksuite/block-suite-editor';
 import * as styles from './page-detail-editor.css';
 
 declare global {
@@ -50,36 +47,19 @@ export const PageDetailEditor = ({ onLoad }: PageDetailEditorProps) => {
     ? pageWidth === 'fullWidth'
     : settings.fullWidthLayout;
 
-  const value = useMemo(() => {
-    const fontStyle = fontStyleOptions.find(
-      option => option.key === settings.fontFamily
-    );
-    if (!fontStyle) {
-      return cssVar('fontSansFamily');
-    }
-    const customFontFamily = settings.customFontFamily;
-
-    return customFontFamily && fontStyle.key === 'Custom'
-      ? `${customFontFamily}, ${fontStyle.value}`
-      : fontStyle.value;
-  }, [settings.customFontFamily, settings.fontFamily]);
-
   return (
-    <Editor
-      className={clsx(styles.editor, {
-        'full-screen': !isSharedMode && fullWidthLayout,
-        'is-public': isSharedMode,
-      })}
-      style={
-        {
-          '--affine-font-family': value,
-        } as CSSProperties
-      }
-      mode={mode}
-      defaultOpenProperty={defaultOpenProperty}
-      page={editor.doc.blockSuiteDoc}
-      shared={isSharedMode}
-      onEditorReady={onLoad}
-    />
+    <CustomEditorWrapper>
+      <Editor
+        className={clsx(styles.editor, {
+          'full-screen': !isSharedMode && fullWidthLayout,
+          'is-public': isSharedMode,
+        })}
+        mode={mode}
+        defaultOpenProperty={defaultOpenProperty}
+        page={editor.doc.blockSuiteDoc}
+        shared={isSharedMode}
+        onEditorReady={onLoad}
+      />
+    </CustomEditorWrapper>
   );
 };
